Stop accidental free of referenced FLAC vendor string
The libFLAC API is unclear about how it handles empty strings. The
original code assumed that empty string would be copied, when 'copy'
param of set_vendor_string was true, as indicated in the API docs.

Checking into the libFLAC source, when a string of length 0 is provided,
it will NOT be copied.  Therfore, free'ing empty string will cause a
SIGSEGV or worse when the FLAC API tries to read the vendor string.