cddb: Plug some memory leaks.
Some of these functions are way, way too big and should be split. One
obvious low hanging fruit is to use some intelligent goto's to clean up and
avoid a bunch of if's in the code, in the very same fashion that is used in
the code of the kernel.

And, to be honest, using some of the coding style conventions of the kernel
here would not hurt at all.