cddb: avoid calling strlen many times in just a few statements.
As strlen is a potentially costly function, it is better if we make a frugal
use of it.

To really make things cleaner here, we should replace all the `\n` or `\r`
with null chars (i.e., `\0`) in just one go. This would reduce the
complexity of the code from something quadratic to something linear in the
length of the string.

While we are at it, `fgets` does not return `EOF` (`fgetc` does, but `fgets`
doesn't) when the end of file is reached and nothing can be read: it returns
`NULL` instead.