Add "Album Artist" field to the list of potential fields to edit
Quoth Daniel (Random Mister):

"I just found an extended patch which also includes support for the
album-artist tag in the fill tags dialog. It is an update of the
other, so if you only add this one to the git tree it should work.
I've been using it for over half a year without bugs, so I think it
can be merged into the Git tree."

Written-By: Daniel <random.mister@gmail.com>