Add option to make genre tags as "Text Only"
Quoth Daniel:

"I also have a patch which sets the genre tag as a string instead of
the genre number. This number is not probably handled by iTunes (it
shows the number, not the actual genre name). I've added it as a
attachment."

This is controlled by an option in the ID3 preferences tab

Written-By: Daniel <random.mister@gmail.com>