Improve album artist entry focus chain order
As reported by Julian Orth <ju.orth@gmail.com>, the album artist entry
was not in the focus chain after the album entry, which is where it
should logically be if looking at the user interface. Reorder the focus
chain so that the order matches the interface layout.

Additionally, use g_list_prepend() to add new focus chain widgets at the
start of the list, and then reverse the list before setting the chain on
the parent container, as this is more efficient.