Use a toolbar for the picture tab buttons
Use a toolbar for the control buttons underneath the picture tab tree
view. Update the sensitivity of the toolbar buttons depending on the
selected items (make the remove, save and property buttons insensitive
if no items are selected). Improve the string for removing images.
Remove some commented-out code for setting the tag area widgets to be
insensitive.