Improve select all action handler, bug 697515
As the action handler is not invoked on a widget, determine the
currently-focused widget and call the appropriate "select all" method to
ensure that Ctrl+A functions correctly in widgets other than the file
list view. Adjust the strings to indicate that the select all action is
not specific to files.