Update the file property sort order automatically
When the sort order is selected in the file menu, and is a property of
the file (rather than a tag property), the file list sort order is not
automatically updated.

This patch makes the file list sort automatically after setting the sort
order in the file menu. Fixes bug 698378.