Refactor list handling in browser.c
Do not check if a GList has a length of 0. Use g_list_free_full() to
free lists with dynamically-allocated elements. Use for loops to iterate
over GList. Store a pointer to the start of the list so that it can be
freed later. Prepend to a list, rather than appending, when populating
it.