Refactor list handling in cddb.c
Do not check if a GList has 0 length. Use a for loop to iterate over
GList. Use g_list_free_full() to free a list with dynamically-allocated
elements. Use g_list_next() rather than GList->next to access the next
element in a list. Keep a pointer to the start of the list so that it
can be freed later. Populate GList by prepending to it, rather than
appending.