Refactor list handling in misc.c
Use a for loop to iterate over GList. Use g_list_next() to access the
next list element, rather than GList->next. Use g_list_free_full() to
free lists with dynamically-allocated elements. Do not check wither a
GList has a length of 0. Keep a pointer to the start of a list so that
it can be freed later.