Refactor list handling in scan.c
Use a for loop to iterate over GList. Use g_list_next() rather than
GList->next to access the next list element. Keep a pointer to the start
of the list so that it can be freed later. Do not check if the length of
a list is 0. Use g_list_free_full() to free lists with
dynamically-allocated elements.