Unref GtkListStores and a GtkTreeStore
GtkListStore and GtkTreeStore inherit from GObject and not
GInitiallyUnowned, and so start with a reference count of 1 and no
floating reference. Therefore, they must be unreffed after adding to a
view in order to be finalized when the view is destroyed.

The remaining list stores are difficult to disentangle from the
surrounding code, so have been left unchanged.