Use correct platform-specific accelarator modifier
GTK+ 2.24.7 (and GTK+ 3.2.1) introduced the concept of a
platform-agnostic accelerator modifier, "Primary", which can be used
instead of "Control". Using "Primary", the correct modifier is chosen
depending on the platform, for example the Option key would be used on
Mac OS X.

Use the new "Primary" modifier and bump the GTK+ dependencies to build
with versions which include the support.