Revert "[patch] debian/patches/01-autoload.patch"
This reverts commit 820c5ae2ddc319073d23d308204069850c1682a2.

A bug where EasyTAG would not load a directory passed as an argument was
fixed and released in version 2.1.7:

https://bugs.launchpad.net/ubuntu/+source/easytag/+bug/230873

However, the underlying cause was an invalid read when checking if the
current path in the browser was valid, in Browser_Update_Current_Path.
Fixing the underlying cause exposed a problem introduced by the original
patch, which lead to the initial directory being loaded twice.

Fixes bug 689445.