Fix localization by defining ENABLE_NLS correctly
Since the switch to use intltool and not gettext in commit
98065d10b09bb27a2dee1e79f9b0eef0cee549a0, localization hsa been broken,
as ENABLE_NLS is not defined by IT_PROG_INTLTOOL but is used in an ifdef
in main().

Additionally, includes of <glib/gi18n-lib.h> were corrected to
<glib/gi18n.h>. Finally, gtk_action_group_set_translation_domain() was
set on action groups, to ensure translations of actions.

Fixes bug 712250.