Return a gboolean from delete-event handler
Avoid undefined behaviour (manifesting as a crash on Fedora Rawhide) by
correctly returning a gboolean in the main window delete-event handler.
If the delete event was propagated by accident, the default handler
would destroy the window and lead to many subsequent problems.

Thanks to Igor Gnatenko for the bug report.

https://bugzilla.gnome.org/show_bug.cgi?id=725568
https://bugzilla.redhat.com/show_bug.cgi?id=1071563