Fixed a album/artist view memory leak
In artist/album view the album list store contains a row for a
collection of all albums for an artist - the "<All Albums>" row. The
attached GList was not freed.

Add a column with a boolean value to the model to be able to identify
this row, as a simple string comparison seems unsafe.