Fix stripping disc number field from ID3v2.4 tags
As part of a change made in bug 726467, disc number handling for ID3
tags was modified to only allow digits and the '/' character in the TPOS
field. A bogus check was added for ID3 tags, so that if saving a tag
with no content in the disc number field, the existing TPOS tag would
not be cleared.

Remove the erroneous check, as et_id3tag_get_tpos_from_file_tag() safely
handles the disc number field being empty.

https://bugzilla.gnome.org/show_bug.cgi?id=742082