Correctly update the artwork description
An EtPicture was fetched from the model containing all the images for a
file, and the description was updated depending on what the user entered
into the artwork properties dialog.

However, the updated EtPicture (which was a copy of the one in the
model, as EtPicture is a boxed type and was fetched from the model with
gtk_tree_model_get()) was not set on the model, and so any change to the
description was lost on exiting the dialog.

Include the EtPicture as part of the gtk_list_store_set() call when
closing the dialog, and update the description correctly.