Add a function for comparing UTF-8 strings
Comparing strings for equality with g_utf8_collate() gives a poor
result, as it is intended for ordering, not equality comparisons. It is
preferable to normalize the string first, and then to compare the
normalized result.

Add a new et_normalized_strcmp0() function to do this, and use it in
et_file_tag_detect_difference() to avoid problems where two strings
compared identically with g_utf8_collate(), which prevented the user
from changing tag fields. Adapt the File_Tag test as appropriate.

https://bugzilla.gnome.org/show_bug.cgi?id=744897