Convert most g_utf8_strlen() to new et_str_empty()
Nearly all uses of g_utf8_strlen() were to check that a string was not
empty. The check for empty was often paired with a check that the char
pointer was not NULL. Simplify the combination of checks into a new
et_str_empty() function, which returns TRUE if either the supplied char
pointer is NULL or the string is empty.

Leave checks that are only concerned with one of the two checks as they
are, but remove g_utf8_strlen() where possible.