Replace filename characters after substitutions
Ensure that scanner options do not override whether illegal characters
should be replaced in filenames, by doing the replacement after the
substitutions. This has the effect that if underscores are to be replaced
by spaces, and the filename has a trailing underscore, the end result
will be a filename with an underscore.

https://bugzilla.gnome.org/show_bug.cgi?id=760155