Use full path when converting name for display
Previously, the full path of a directory was converted to UTF-8 for
display in the UI. Commit 9d23932754a9b82db0b1be3ac02df04717e99a6c
changed the behaviour to use the display name from GFileInfo, which only
contains the basename of the file and not the complete path. Use the
full path, and convert it for display using g_filename_display_name().