Avoid locale-sensitive field name reading for Ogg
Avoid calling strncasecmp() when searching for tag field names in Vorbis
comments while reading Ogg (Vorbis and Opus) files. The same function is called
internally by libvorbis when searching for tags, and it is dependent on the
current locale, so it fails with Turkish and the lower-case 'i' being
transformed to the dotted upper-case version. As Vorbis comment field names are
only permitted to be ASCII, with some further restrictions, use g_ascii_strup()
to convert them all to upper-case before comparing them.

As an additional optimization, insert the tag key/value pairs in a hash table
during reading. The avoids the behaviour of searching through all the comments
in the file when loading data for each UI field (as well as a second time, when
populating the list of unsupported fields). instead, only a single lookup per
tag is required.

https://trac.xiph.org/ticket/2079