Remove setlocale() calls around tag reading
It is unsafe to call setlocale() in a multi-threaded program (and EasyTAG is,
because GLib creates a worker thread for handling D-Bus calls). Additionally,
it is no longer necessary to change the locale when reading files, as all
locale-sensitive tag reading code has been converted to be insensitive to the
current locale.