Avoid crash when saving audioless ID3v2.4 tags
In the case where an ID3v2.4 tag was written that differed in size to
the previous tag, audio was assumed to exist after the tag. This
assumption failed if the file consisted solely of the ID3 tag and
nothing else. Avoid the crash by skipping the audio reading and writing
if there is none to copy.

https://mail.gnome.org/archives/easytag-list/2016-June/msg00002.html