Fix Ogg tag writing on Windows
Under Windows, if a file is opened for reading, it cannot be opened a
second time for writing. The Vorbis comment code keeps the file open
for reading throughout the process of writing new tags, mostly because
the audio stream must be read into memory before being written out
after the new tags. Avoid the problem by closing the input stream only
during the writing process, and at all other times maintaining a
consistent state, with the input stream open.

https://mail.gnome.org/archives/easytag-list/2016-November/msg00000.html