Fix preserving of non-UI tag fields
After refactoring of some of the tag copying functions in
9aad89da3bbe7c822eb12fb082e7d4a085eb3ed6, "other" tag fields (those not
shown in the UI, but intended to be preserved) were lost when saving a
file with modified tags. There was one instance in the refactoring where
arguments to et_file_tag_copy_other_into() should have been swapped, and
were not.

Thanks to Philippe Troin for the bug report and initial patch.

https://bugzilla.gnome.org/show_bug.cgi?id=795018