Use a GtkHeaderBar instead of a toolbar.
The headerbar includes:
* "show scanner" button
* "stop ongoing operation" button
* "undo"/"redo" buttons

* radiobuttons for switching between the 2 views

* "find" button
* save button
* primary menu

The primary menu contains actions which are not in the headerbar, and which are not related to the "browser" component. Some actions might be polemic, like the "load filenames from a text file" whch is *very* specific. Some labels (like the undo/redo items, or the "cddb" thing) are improvable imo.