Don't set the appmenu if the DE doesn't want it
The rationale here is that easytag will probably be used as a flatpak app on Ubuntu 18.04 or Debian Buster for a bunch of years. These users should have a consistent UX, and the 10 lines of C giving them the appmenu isn't a big deal.

This commit also hide the menubar, which is now available only for people with a "global menu" in their DE