Use C++ linker to check for id3lib
Instead of manually adding -lstdc++ to the linker command line, which
isn't going to work if the C++ runtime library isn't provided by GCC,
use the C++ compiler driver to link the program to avoid the need to
figure out the name of the C++ runtime library ourselves.

This fixes the check on FreeBSD, which uses Clang and LLVM libc++ by
default.